tests: Add pthread based concurrency stress test for babl_fish()
Add a test that runs a bunch of pthread threads, all trying to use
babl_fish() at the same time. This test currently fails and it is the
goal of bug 587675 to make it pass.

The cause of the failure is the static variables in babl-fish-path.c.
In particular 'current_path' clearly needs to be moved to some
instance variable.