Only initialize test buffers once
Also made the test buffer static inline for individual conversions like
it already was for fish paths.

Note, this relies on none of the tested conversions by accident
being buggy enough to modify their source data, adding a sanity
function that compares the buffers with the correct values could
perhaps be useful.