Add code to handle N-components
Added a new API call, babl_format_n (). This allows constructing a babl
format with an arbitrary number of components all of the same type. At
the moment. The model used for such formats is "Y", at the moment
conversions are only well defined for such formats to other n-component
formats with the same number of components.