Add BABL_IS_BEING_COMPILED #define support
Sometimes it is ok that babl includes are included directly and not
through babl.h, like when we generate GObject introspection files.

The important thing is just that users of the babl API don't begin to
include headers we don't guarantee to always exist, and
BABL_IS_BEING_COMPILED doesn't obstruct this goal.