Minimal float -> u8, u16 conversions
Add some simplistic but "accurate" float -> u8 and float -> u16
conversions to make more non-reference paths usable.

The main difference from the existing conversions is that these
clamp out of range values and make the path error checker happy.

I also had to fudge the float-to-8bit slightly because the prior
values cause it to check for the rounding on an exact 0.5 value.