Increase the precision of ref gamma conversions
Use a 3 round Newtons method for the fast version. This should be
fully accurate for floats in [0, 1], less so for very small values
but they're in the the linear part of the sRGB curve.

For the reference conversions the standard C pow() function is
used, which give accuracy far in excess of anything we need.