tools: add babl-icc-dump
This commandline tool contains code to parse the v4 elements babl-icc currently
doesn't make use of. It loads data from the ICC profile directly into double
precision floats.

  CIE xy red: 0.640009 0.330022
  CIE xy green: 0.300008 0.600015
  CIE xy blue: 0.150011 0.059998

Is due to IEEE double precision floats not being as precise as the fixed point
values in the profile, which are *exactly*:

  CIE xy red:   0.6400  0.3300
  CIE xy green: 0.3000  0.6000
  CIE xy blue:  0.1500  0.0600