babl-fish-reference: implement floating point version of reference path
This relies on conversions, linear or planar, being registered to and from
RGBA float, similar to the already existing reference conversions in double.

To enable use of it set the environment variable BABL_REFERENCE_FLOAT to any
value, babl will crash if the above requirement is not met for a requested
conversion - permitting adding it - fallback code could be added that implements
it using the double versions when missing - but not having a full coverage at
least for the commonly used formats defeats the purpose.