babl: reduce default max path length to 3 (5)
With the space invasion progressing, we're starting to have
many more conversions to search through - in some cases
looking for conversions - when there are no conversions
become too expensive with an exhaustive search of all
conversions up to 7 steps. (it is likely possible to reduce
the amount of conversions tried).

As well re-enable handling of grayscale ICC profiles - since
the formerly patological tests on gimp now pass.