meson.build: Fix lcms2 detection for MSVC builds
The build files for lcms2 do not normally generate pkg-config files for us, so
we need to look for it manually if they cannot be found.  Also, at least for
32-bit builds, we need to check whether the found lcms2 .lib is really a DLL
build or a static .lib build, otherwise linking will fail for DLL lcms2 builds
if we do not define CMS_DLL for our builds, at least for 32-bit builds.