extensions: Export init() on MSVC builds
We also do not want to use GCC-esque linker flags which MSVC does not
understand.

All tests on Visual Studio builds pass without the SIMD optimizations at this
point, as they are being disabled, which will be enabled later.

Visual Studio 2017 15.5 or later will be required due to the use of musl
items.