babl: babl-fish-path improve portability for ppc
Quoting his explaination for why this fixes things (the other half of
the same function was already portable.)

"This breaks on PowerPC 32-bit, because the calling convention for
passing a union is to pass a pointer to a temporary copy of the union.
This pointer isn't as a function pointer.  On some other platforms, the
call would just copy the union, which is like copying the function
pointer.  On all platforms, the compiler doesn't check the type, because
babl_conversion_new() has a va_arg(3) prototype."

Portabilitiy issue figured out by George Koehler, This fixes issue #24.