ci: Unbreak archlinux-based CI
Recently Arch has changed its glibc package to require kernel
4.4[1] as a poor man's attempt to disable faccessat2 syscall that
is known to cause problems with old libseccomp and Docker. Downgrade
glibc to the pkgrel before the change and ensure it's not upgraded.

Additionally switch to the official Arch Linux image in the library
namespace. The base image has been deprecated.

[1] https://github.com/archlinux/svntogit-packages/commit/893b1c268abc8822332655865e3d4546025a9b4b