bin: new "babl" tool to convert pixel values in command line.
This is a CLI tool to quickly make conversions of raw pixel values from
one known format to another.

This is a first basic version, which supports input and output in float,
u8, u16 and u32 (no "half" type yet in particular).

Also it doesn't support profiles yet, or rendering intents, but soon
will.

Finally it doesn't have a "help" output yet to help usage.