Add a few babl libc wrappers
Add the following wrappers:

* babl_fopen
* babl_rename
* babl_remove
* babl_mkdir
* babl_stat

And the following util:

* babl_dir_list

On Windows, the wrappers internally convert from UTF8 to UTF16 and
call the wide char routines, that way the string need not be limited
to the ANSI character set. Outside of Windows the wrappers are no-op.
babl_dir_list is implemented using opendir / readdir on non-Windows
and _wfindfirst64 / _wfindnext64 on Windows.