babl: limit LUT creation for 4→8 and 4→16 bpp further.
Looking at LUT conversion code, we are clearly only supporting
converting from "u8" (to "u16" or "float" respectively).

Conversions such as "YA half" to "R'G'B'A float" for instance would be
completely broken (it was trying to use only one byte of the "YA half",
assumed it was "u8" and transform it to "float"). This is why the sample
image in #85 was appearing on display as more transparent than it should
when converted to 16-bit floating point grayscale image.