babl, bin: support "half" type in the babl CLI tool.
The code to encode half values into and from float values was moved from
type-half.c to babl-util.h as an internal function, to make it usable by
the CLI code.

It may not seem totally useful to have half support since anyway we
convert from/to string for the CLI tool (in this regard, half and float
are the same). But since the tool performs an actual babl conversion,
it's actually a good way to test quickly that half conversion works
properly too.