bin: fix linking for cross-compiled Windows build.
While I didn't have a problem for Linux builds, when cross-compiling for
Windows, these errors occur:

> babl-util.c:296: undefined reference to `babl_strcat'

(and a few similar other "undefined reference" as this one gets fixed)

This is because we include babl-internal.h in bin/babl.c but don't add
the right compilation unit sources. The build was still finding them (or
maybe ignoring them as the functions were unused in the babl CLI tool
code) on Linux, but not on Windows.