babl, babl: new babl-shared-util.[ch] for functions shared between CLI and lib.
While apparently fixing the compilation for Windows (and still compiling
on Linux), my previous commit was in fact crashing the CLI tool when it
runs babl_init() in some infinite loop (it looks like it's crashing
inside babl logging code, hence the infinite looping).

Anyway this was annoying me that we had to build-in more than necessary
from babl-internal when all I wanted was to share 2 functions. So I
split babl-util with a new babl-shared-util. The latter is to be used
only for the few utility functions which we might want to share with the
CLI tool (so far, only the 2 half to/from float (en|de)coders).

This time, I've not only compiled but also run the CLI tool both on
Linux and with Wine for the Windows cross-build.