bin: add "double" support to the CLI tool.
Again, even though it's basically the same as float or half for a CLI
tool (where all these floating point values are just converted from
string representation), it is still interesting since the tool is doing
actual conversion with a fish using the requested formats. Therefore
it's a nice way to quickly test conversions in terminal.