extensions: component "alpha" should be registered with "alpha" attribute.
Though the models may itself be registered with a similarly named
attribute to mean they encode alpha, babl_format_has_alpha() in
particular works by verifying if any of the components
represents/encodes alpha. As a consequence, the function was failing to
consider e.g. HSVA as a model encoding alpha.