Issue #101: "RGBA float" to "CIE Lab alpha float" fish with SSE2…
… codepath give different results than without SSE2.

The slightly different result is expected and the SSE2 code does not
seem bugged at all. The problem though is that in a single babl_process,
we end up with different results, which may be amplified by further
algorithm into surprising renders (cf. gimp#12030).
With this commit, we ensure than when using the SSE2 codepath, we do not
end processing the last few pixels with different code.