extensions: always use the SSE2 codepath when the processor has support.
rgbaf_to_Labaf_sse2() was falling back to the non-SSE2 algorithm when
input and output were not 16-byte-aligned. Additionally to missing some
good optimization, the slightly different implementation makes for
creating opportunities for artifact-looking results when a single canvas
is broken into smaller parts, some of which may be processed with one
implementation, others with the other implementation.

See:
https://gitlab.gnome.org/GNOME/gimp/-/issues/12030
https://gitlab.gnome.org/GNOME/gimp/-/issues/2933