bin: add ability to select one of the known spaces by name in CLI tool.
--input-profile and --output-profile options are replaced respectively
by --input-space and --output-space (but the old option names still work
for scripts compatibility).

They can still be used to specify an ICC file like we used to, but
additionally if you start with '*', you can select one of the known
spaces, such as sRGB, Rec2020, etc.

In the weird case that you wanted to select a profile in a path starting
with '*' (it's not a forbidden character by the file system even though
usually used for glob file selection), you can protect the initial
character with a double asterisk '**'.