extensions: add direct conversions from "R'G'B'A" to/from "HSL".
On commonly-sized images, the HSL conversions went from near instant to
"seeing-the-tiles-slowly-updating" after we switched some GIMP code to
babl conversion.

This is because the code is actually made to work in gamma 2.2 and the
only conversion was from and to linear RGB. So when converting
non-linear RGB, we were making far too many conversions. Let's have
faster conversion paths.