babl: fix broken trc cache for formula types != gamma
The matching of pre-existing TRCs only worked well for the gamma type of
formulas, other formulas would get the first registered trc of its type. These
are not our fastest conversions - but they will be baked into LUTs if
frequently used and the source has <24bit of color components.

The generated names used contain the parameters of the requested trcs,
thus moving it within the range used for searching makes us return
the correct result.

Fixes #66