babl: space and reference handling of black point compensation and intents
We Only do BPC if both source and destination spaces intent have BPC set, BPC being
set is the default.

This logic extends to other deviations from the BABL_ICC_INTENT_DEFAULT as well, the
non-default of a pair wins - if both are non-default the target space wins.

This is so that having buffers with default intents do nothing while having an
out-of-the ordinary buffer only affects the intent of that buffer, while
preserving the ability of babl formats to roundtrip data.

Still missing : cmyk <-> cmyk - it is always done as relative colorimetric with bpc