tests: move trcs test to platform_unix
The recent addition of support for Visual Studio on Windows, and more
specifically adding `vs_module_defs` to the library definition in
meson, means that private symbols can't be accessed anymore outside
the library.

The trcs is trying to call `babl_trc_new`, which is a private symbol,
and on Windows this now causes a crash as soon as that function is
called.
Since this only affects this test on Windows, let's set that test as
Unix only.