babl: fallback to build-time BABL_PATH when our relocatable heuristic fails.
This may happen when babl is linked by a build-time only tool, in which
case it won't be installed under bin/ of the destination prefix.
In particular we had the case when introspecting libgimp:

> FAILED: libgimp/Gimp-3.0.gir
> […]
> ../babl/babl-internal.h:218 babl_fatal()
> 	Relocatable builds require the executable to be installed in bin/ unlike: /builds/GNOME/gimp/_build-aarch64/tmp-introspectgu6s05db/Gimp-3.0

So we just fall back to the build-time path, which assumes of course
that we didn't relocate babl already. This is a pretty safe assumption
since you should not move your prefix all the time while building a full
bundle with many dependencies.