babl: allow relocatable binaries to be under bin/ or lib*/ at any level.
So I had forgotten that plug-ins will also run babl and they are not
under bin/, but under the lib*/ hierarchy.

E.g.: $PREFIX/lib64/gimp/3.0/plug-ins/file-png/file-png

So I not only check for a parent directory starting with 'lib', but I
also go as deep in the hierarchy as needed. This will work out for
plug-ins as well as take care of the multi-arch folder hierarchy.